Found this great video on how to setup a simple camera:
https://www.youtube.com/watch?v=ceBCDKU_mNw

Used this post:

https://stackoverflow.com/questions/20483809/c-sharp-xna-4-0-camera-zoom-and-sprite-resizing

To learn that you can do zooms using the matrix transformation Matrix.CreateScale. Its also possible to do rotations just by multiplying the final transformation with a CreateRotation. The transformation can just be given to the spritebatch begin function to apply.

The Render target is now used to draw static things to the screen like health bars or debug text or status items and stuff like that. It is also now being used primarly for scaling to a fixed screensize as it should be rather than trying to use it as a camera.